Add error handling and fluent validation

+++ SportyBuddies.Application/Sports/Commands/DeleteSport/DeleteSportCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using SportyBuddies.Application.Common.Interfaces;
+using SportyBuddies.Application.Exceptions;
 
 namespace SportyBuddies.Application.Sports.Commands.DeleteSport;
 
@@ -17,6 +18,8 @@ public class DeleteSportCommandHandler : IRequestHandler<DeleteSportCommand>
     public async Task Handle(DeleteSportCommand command, CancellationToken cancellationToken)
     {
         var sport = await _sportsRepository.GetByIdAsync(command.SportId);
+
+        if (sport == null) throw new NotFoundException(nameof(sport), command.SportId.ToString());
 
         await _sportsRepository.RemoveSportAsync(sport);
         await _unitOfWork.CommitChangesAsync();
